requirements.txt
Flask==2.1.1
Pillow==9.1.0
qrcode==7.3.1
Werkzeug==2.1.1
gunicorn==20.1.0